backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
chromadb==0.4.18
numpy==1.24.3
requests==2.31.0
email-validator==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
pymongo==4.6.1
motor==3.3.2
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
google-generativeai==0.3.2
sendgrid==6.12.5